actualizacion precio de vehiculos

+++ Rebits/database/migrations/2024_01_18_230952_create_vehiculos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('modelo');
             $table->unsignedSmallInteger('anho');
             $table->foreignId('duenho_id')->constrained('usuarios');
-            $table->decimal('precio', 10, 2);
+            $table->integer('precio');
             $table->timestamps();
         });
     }
